% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plasticity.R
\encoding{UTF-8}
\name{A.line}
\alias{A.line}
\title{A-Line (Plasticity)}
\usage{
A.line(LL)
}
\arguments{
\item{LL}{= Liquid Limit (percent)}
}
\value{
The plasticity index corresponding to the A-line on Casagrande's
   plasticity chart
}
\description{
This function calculates the A-line on Casagrande's plasticity chart. 1
December 2015
}
\note{
Either PL or PI must be specified
}
\author{
Kyle Elmy and Jim Kaklamanos
}
