% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilClassification.R
\encoding{UTF-8}
\name{AASHTO}
\alias{AASHTO}
\title{AASHTO Soil Classification (AASHTO Group Classification)}
\usage{
AASHTO(
  sieve = NA,
  size = NA,
  percent = NA,
  metric = NA,
  p10 = NA,
  p40 = NA,
  p200 = NA,
  LL = NA,
  PL = NA,
  PI = NA,
  NP = NA
)
}
\arguments{
\item{sieve}{= vector of sieve numbers (according to ASTM D422) in
grain-size distribution [Grain-size parameter]}

\item{size}{= vector of grain sizes (in or mm) in distribution [Grain-size
parameter]}

\item{percent}{= vector of percent passing in grain-size distribution
[Grain-size parameter]}

\item{metric}{= logical variable for grain-size distribution: TRUE for
metric units (mm), FALSE for English units (in); only required if "size" is
supplied [Grain-size parameter]}

\item{p10}{= percent passing #10 sieve [Grain-size parameter]}

\item{p40}{= percent passing #40 sieve [Grain-size parameter]}

\item{p200}{= percent passing #200 sieve [Grain-size parameter]}

\item{LL}{= liquid limit (percent) [Plasticity parameter]}

\item{PL}{= plastic limit (percent) [Plasticity parameter]}

\item{PI}{= plasticity index (percent) [Plasticity parameter]}

\item{NP}{= logical variable indicating whether the soil is nonplastic (TRUE
or FALSE) [Plasticity parameter]}
}
\value{
AASHTO group classification
}
\description{
This set of functions classifies soil using the American Association of
State Highway and Transportation Officials (AASHTO) Soil Classification
System. Inorganic soil classification using the AASHTO method. 1 December 2015
}
\note{
For the grain-size data, one of three following pieces of data must be
specified:
1.  Sieve numbers (sieve); and percent passing
2.  Grain sizes (size); and percent passing
3.  p10, p40, p200
If sieve data are specified, for sieves larger than the no. 4 sieve,
the user should specify the sieve size in inches 
(e.g., 3/8, 3/4, 1, 1.5, 2, 3, etc.)
If sieve or size data are specified, this function assumes that the
no. 10, 40, and 200 sieves have been used.
For the plasticity data, one of three following pieces of data must be
specified:
1.  LL and PL
2.  LL and PI
3.  NP (if the soil is nonplastic)
To obtain the group index using the AASHTO method, use the GI function.
}
\examples{
AASHTO(p10 = 51, p40 = 30, p200 = 15, LL = 40, PI = 10)
}
\author{
Kyle Elmy and Jim Kaklamanos
}
