% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearingCapacity.R
\encoding{UTF-8}
\name{bearingCapacity}
\alias{bearingCapacity}
\title{Ultimate Bearing Capacity (Terzaghi)}
\usage{
bearingCapacity(
  phi,
  c,
  B,
  L,
  D,
  Dw,
  gamma,
  gammaW = NA,
  metric,
  case = "general",
  shape = "square"
)
}
\arguments{
\item{phi}{= effective friction angle (deg)}

\item{c}{= effective cohesion (psf or kPa)}

\item{B}{= foundation width (ft or m), or foundation diameter for circular
footings}

\item{L}{= foundation length (ft or m)}

\item{D}{= Depth of foundation (ft or m)}

\item{Dw}{= Depth of groundwater table below foundation base (ft or m)}

\item{gamma}{= unit weight of soil (pcf of kN/m^3)}

\item{gammaW}{= unit weight of water (default = 62.4 pcf for English units;
9.81 kN/m^3 for metric units)}

\item{metric}{= logical variable: TRUE (for metric units) or FALSE (for
English units)}

\item{case}{= "general" or "local" to indicate general or local shear
failure ("general" is default)}

\item{shape}{= "square", "rectangle", "circle", "strip" (or "continuous")}
}
\value{
Bearing capacity (q_ult) from Terzaghi's simple theory (psf or kPa)
}
\description{
This function calculates the ultimate bearing capacity from Terzaghi's simple
theory. 1 October 2015
}
\note{
Either SI or English units can be used, but must stay consistent.
When specifying the length and width of rectangular foundations, L should be
the longer of the two lengths.             
When the groundwater table is deep or unknown, set Dw >= D + B
}
\author{
Kyle Elmy and Jim Kaklamanos
}
