% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrainSize.R
\encoding{UTF-8}
\name{Dsize}
\alias{Dsize}
\title{D-Size (Grain-Size Distributions)}
\usage{
Dsize(N, sieve = NA, size = NA, percent, metric)
}
\arguments{
\item{N}{= the percent corresponding to the desired D-size}

\item{sieve}{= vector of sieve numbers (according to ASTM D422) that make up
the grain-size distribution}

\item{size}{= vector of grain sizes (in or mm) of the distribution}

\item{percent}{= vector of percent passing of the grain-size distribution}

\item{metric}{= logical variable: TRUE for metric units (mm), FALSE for
English units (in)}
}
\value{
The grain size corresponding to a certain percent finer (N), given a
    grain-size distribution
}
\description{
This function calculates the diameter size. 1 December 2015
}
\note{
For sieves larger than the no. 4 sieve, the user should specify the sieve size
in inches (e.g., 3/8, 3/4, 1, 1.5, 2, 3, etc.)
Either sieve numbers OR grain sizes must be provided.  The "metric" variable is
only required if sieve numbers are provided.
The function uses logarithmic interpolation to calculate the D-size from the
provided grain-size distribution
Log-linear extrapolation is used for grain sizes beyond the range of the data,
and a warning is provided
}
\examples{
sieve.example <- c(3/8, 4, 10, 20, 40, 140, 200)
percent.example <- c(95.72, 90.23, 81.49, 66.36, 50.00, 8.51, 4.82) 
Dsize(N = 50, sieve = sieve.example, percent = percent.example, metric = TRUE)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
