% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlopeStability.R
\encoding{UTF-8}
\name{FSinf}
\alias{FSinf}
\title{Slope Stability (Infinite Slope Analysis)}
\usage{
FSinf(c, phi, gamma, gammaW = NA, alpha, D, zw, metric)
}
\arguments{
\item{c}{= soil cohesion [Soil parameter]}

\item{phi}{= soil friction angle (degrees) [Soil parameter]}

\item{gamma}{= soil unit weight [Soil parameter]}

\item{gammaW}{= unit weight of water (default = 62.4 pcf for English units;
9.81 kN/m^3 for metric units) [Soil parameter]}

\item{alpha}{= slope angle (degrees) [Slope geometry]}

\item{D}{= depth to failure plane [Slope geometry]}

\item{zw}{= distance of groundwater table above failure plane
(use 0 for a dry slope and D for a submerged slope with parallel seepage)
[Slope geometry]}

\item{metric}{= logical variable: TRUE (for metric units: kN/m^3) or FALSE
(for English units: pcf) [this is needed if gammaW is unspecified] [Units]}
}
\value{
Factor of safety against shear failure on slopes using infinite
    slope analyses
}
\description{
This function determines the slope stability using the infinite slope
analysis. 1 January 2016
}
\note{
Assumptions of infinite slope analyses include (Coduto et al., 2011):
1. The slope face is planar and of infinite extent.
2. The failure surface is parallel to the slope face.
3. Vertical columns of equal dimensions through the slope are identical.
}
\author{
Jim Kaklamanos and Kyle Elmy
}
