% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlopeStability.R
\encoding{UTF-8}
\name{FSplanar}
\alias{FSplanar}
\title{Slope Stability (Planar Failure Analysis)}
\usage{
FSplanar(c, phi, alpha, L, W, u)
}
\arguments{
\item{c}{= soil cohesion [Soil parameters]}

\item{phi}{= soil friction angle [Soil parameters]}

\item{alpha}{= angle of failure plane (degrees) [Slope geometry]}

\item{L}{= length of failure plane [Slope geometry]}

\item{W}{= weight of slope per unit width [Slope geometry]}

\item{u}{= average pressure head on the failure plane [Slope geometry]}
}
\value{
Factor of safety against shear failure on slopes with a planar
    failure surface
}
\description{
This function determines the slope stability using the planar failure
analysis. 1 January 2016
}
\note{
Either English or metric units can be used, but they must be consistent.
}
\author{
Jim Kaklamanos and Kyle Elmy
}
