% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotech-package.R
\docType{package}
\name{geotech-package}
\alias{geotech}
\alias{geotech-package}
\title{geotech: Geotechnical Engineering}
\description{
geotech is a compilation of functions for performing calculations and
creating plots that commonly arise in geotechnical engineering and soil
mechanics. The types of calculations that are currently included are:
(1) phase diagrams and index parameters, (2) grain-size distributions, 
(3) plasticity, (4) soil classification, (5) compaction, (6) groundwater,
(7) subsurface stresses (geostatic and induced), (8) Mohr circle analyses,
(9) consolidation settlement and rate, (10) shear strength, (11) bearing
capacity, (12) lateral earth pressures, (13) slope stability, and (14)
subsurface explorations. Geotechnical engineering students, educators,
researchers, and practitioners will find this package useful.
}
\keyword{internal}
