% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilClassification.R
\encoding{UTF-8}
\name{GI}
\alias{GI}
\title{AASHTO Group Index}
\usage{
GI(p200, LL, PL = NA, PI = NA)
}
\arguments{
\item{p200}{= fines content, i.e. percent passing the #200 sieve (percent)}

\item{LL}{= liquid limit (percent)}

\item{PL}{= plastic limit (percent)}

\item{PI}{= plasticity index (percent)}
}
\value{
AASHTO group index (rounded to the nearest integer)
}
\description{
Calculate a soil's group index using the AASHTO method. 1 December 2015
}
\note{
Either PL or PI must be specified
}
\examples{
GI(p200 = 48, LL = 45, PI = 21)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
