% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrainSize.R
\encoding{UTF-8}
\name{grainSize.coefs}
\alias{grainSize.coefs}
\title{Coefficients of Uniformity and Curvature (Grain-Size Distributions)}
\usage{
grainSize.coefs(percent, sieve = NA, size = NA, D10 = NA, D30 = NA, D60 = NA)
}
\arguments{
\item{percent}{= vector of percent passing}

\item{sieve}{= vector of sieve numbers according to ASTM D422}

\item{size}{= vector of grain sizes}

\item{D10}{= D-sizes corresponding to 10 percent}

\item{D30}{= D-sizes corresponding to 30 percent}

\item{D60}{= D-sizes corresponding to 60 percent}
}
\value{
A two-element list containing: Cu = Coefficient of uniformity (D60 /
    D10) and Cc = Coefficient of curvature (D30^2 / (D10 * D60))
}
\description{
This function calculates the coefficients of uniformity and curvature. 1
December 2015
}
\note{
For sieves larger than the no. 4 sieve, the user should specify the sieve
size in inches (e.g., 3/8, 3/4, 1, 1.5, 2, 3, etc.)
The user has three options for input to this function:
1.  Sieve numbers (sieve); and percent passing
2.  Grain sizes (size); and percent passing
3.  D10, D30, and D60; and percent passing
}
\examples{

# Example 1

sieve.example <- c(3/8, 4, 10, 20, 40, 140, 200)
percent.example <- c(95.72, 90.23, 81.49, 66.36, 50.00, 8.51, 4.82) 
grainSize.coefs(sieve = sieve.example, percent = percent.example)


# Example 2

grainSize.coefs(D60 = 0.10, D30 = 0.03, D10 = 0.002)


}
\author{
Kyle Elmy and Jim Kaklamanos
}
