% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrainSize.R
\encoding{UTF-8}
\name{grainSize.plot}
\alias{grainSize.plot}
\title{Grain-Size Plot}
\usage{
grainSize.plot(sieve = NA, size = NA, percent, metric)
}
\arguments{
\item{sieve}{= vector of sieve numbers according to ASTM D422}

\item{size}{= vector of grain sizes (in or mm)}

\item{percent}{= vector of percent passing}

\item{metric}{= logical variable: TRUE for metric units (mm), FALSE for
English units (in)}
}
\value{
Plot of soil's grain-size distribution
}
\description{
This function produces the grain-size plot. 1 December 2015
}
\note{
For sieves larger than the no. 4 sieve, the user should specify the sieve
size in inches (e.g., 3/8, 3/4, 1, 1.5, 2, 3, etc.)
Either sieve numbers OR grain sizes must be provided.
}
\examples{
sieve.example <- c(3/8, 4, 10, 20, 40, 140, 200)
percent.example <- c(95.72, 90.23, 81.49, 66.36, 50.00, 8.51, 4.82) 
grainSize.plot(sieve = sieve.example, percent = percent.example, metric =
TRUE)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
