% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stress.R
\encoding{UTF-8}
\name{induced.area.profile}
\alias{induced.area.profile}
\title{Induced Stress Due to Area Load: Profile}
\usage{
induced.area.profile(z = NA, q, B, L = NA, shape)
}
\arguments{
\item{z}{= vector of depths (default: 1-ft or 1-m increments, to a maximum
depth of 50 ft or 50 m)}

\item{q}{= applied pressure at ground surface}

\item{B}{= width of loaded area}

\item{L}{= length of loaded area (rectangular foundations only)}

\item{shape}{= shape of loaded area (a string containing "circle", "square",
"strip", or "rectangle")}
}
\value{
A two-element list: depth = vector of depths and sigmaZ = vector of
   induced vertical stresses below center of the loaded area
}
\description{
This function computes the vertical stress at a point. 1 October 2015
}
\note{
This function currently uses the approximate method of Poulos and Davis (1974).
More advanced formulations are expected in future versions of this package.
}
\examples{
induced.area.profile(q = 1000, B = 3, shape = "square")


}
\author{
Kyle Elmy and Jim Kaklamanos
}
