% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stress.R
\encoding{UTF-8}
\name{induced.area}
\alias{induced.area}
\title{Induced Stress Due to Area Load}
\usage{
induced.area(z, q, B, L = NA, shape)
}
\arguments{
\item{z}{= depth of interest}

\item{q}{= applied pressure at ground surface}

\item{B}{= width of loaded area}

\item{L}{= length of loaded area (rectangular foundations only)}

\item{shape}{= shape of loaded area (a string containing "circle", "square"
   , "strip", or "rectangle")
#'}
}
\value{
Induced vertical stress at the center of the loaded area
}
\description{
This function computes the induced stress due to area load. 1 October 2015
}
\note{
This function currently uses the approximate method of Poulos and Davis (1974).
More advanced formulations are expected in future versions of this package.
}
\examples{
induced.area(z = 10, q = 1000, B = 3, shape = "square")

}
\author{
Kyle Elmy and Jim Kaklamanos
}
