% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stress.R
\encoding{UTF-8}
\name{induced.point.profile}
\alias{induced.point.profile}
\title{Induced Stress Due to Point Load: Profile}
\usage{
induced.point.profile(P, x, y, z = NA, nu)
}
\arguments{
\item{P}{= point load}

\item{x}{= horizontal distance from load in the X direction}

\item{y}{= horizontal distance from load in the Y direction}

\item{z}{= vector of depths of interest (default: 1-ft or 1-m increments,
to a maximum depth of 50 ft or 50 m)}

\item{nu}{= Poisson's ratio}
}
\value{
A seven-element list containing seven vectors displaying the depth
   variation of induced stresses due to a point load following the Boussinesq
   method: depth = vector of depths, sigmaX = induced horizontal stress in the
   x direction, sigmaY = induced horizontal stress in the y direction, sigmaZ
   = induced vertical stress, tauZX = induced shear stress on the ZX plane,
   tauYX = induced shear stress on the YX plane, tauYZ = induced shear stress
   on the YZ plane
}
\description{
This function computes the induced stress due to point load: profile. 1
October 2015
}
\examples{
induced.point.profile(P = 100000, x = 5, y = 2, nu = 0.35)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
