% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stress.R
\encoding{UTF-8}
\name{induced.point}
\alias{induced.point}
\title{Induced Stress Due to Point Load}
\usage{
induced.point(P, x, y, z, nu)
}
\arguments{
\item{P}{= point load}

\item{x}{= horizontal distance from load in the X direction}

\item{y}{= horizontal distance from load in the Y direction}

\item{z}{= depth of interest}

\item{nu}{= Poisson's ratio}
}
\value{
A six-element list containing the induced stresses due to a point
   load following the Boussinesq method: sigmaX = induced horizontal stress in
   the x direction, sigmaY = induced horizontal stress in the y direction,
   sigmaZ = induced vertical stress, tauZX = induced shear stress on the ZX
   plane, tauYX = induced shear stress on the YX plane, tauYZ = induced shear
   stress on the YZ plane
}
\description{
This function computes the induced stress due to point load. 1 October 2015
}
\examples{
induced.point(P = 100000, x = 5, y = 2, z = 6, nu = 0.35)
}
\author{
Kyle Elmy and Jim Kaklamanos
}
