% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LateralEarthPressures.R
\encoding{UTF-8}
\name{K}
\alias{K}
\title{Coefficient of Lateral Earth Pressure (General)}
\usage{
K(sigmax, sigmaz)
}
\arguments{
\item{sigmax}{= Horizontal effective stress}

\item{sigmaz}{= Vertical effective stress}
}
\value{
Coefficient of lateral earth pressure
}
\description{
This function calculates the general coefficient of lateral earth pressure.
1 October 2015
}
\examples{
K(sigmax = 50, sigmaz =90)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
