% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LateralEarthPressures.R
\encoding{UTF-8}
\name{Ka}
\alias{Ka}
\title{Rankine Active Earth Pressures (Lateral Earth Pressure)}
\usage{
Ka(phi, beta = 0)
}
\arguments{
\item{phi}{= effective friction angle (deg)}

\item{beta}{= angle of backfill (deg); default = 0}
}
\value{
Coefficient of lateral earth pressure (active)
}
\description{
This function calculates the active coefficient of lateral earth pressure.
1 October 2015
}
\note{
The Rankine theory requires that beta <= phi
}
\examples{
Ka(phi = 30, beta = 10)



}
\author{
Kyle Elmy and Jim Kaklamanos
}
