% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groundwater.R
\encoding{UTF-8}
\name{kConstant}
\alias{kConstant}
\title{Hydraulic Conductivity From Constant Head Test (Groundwater)}
\usage{
kConstant(V, t, h, L, As = NA, Ds = NA)
}
\arguments{
\item{V}{= volume of water collected}

\item{t}{= time of flow}

\item{h}{= head difference between inflow and outflow}

\item{L}{= length of soil sample}

\item{As}{= cross-sectional area of the soil sample}

\item{Ds}{= diameter of soil sample}
}
\value{
Hydraulic conductivity from the constant head test
}
\description{
Computes the hydraulic conductivity from the constant head test. 1 January
2016
}
\note{
Either English or metric units can be used, but they must be consistent.
Either the area or the diameter of the soil sample need to be specified
}
\examples{
kConstant(V = 800, t = 100, h = 200, As = 40, L = 50)
# k in units of cm/s

}
\author{
Kyle Elmy and Jim Kaklamanos
}
