% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groundwater.R
\encoding{UTF-8}
\name{kFalling}
\alias{kFalling}
\title{Hydraulic Conductivity From Falling Head Test (Groundwater)}
\usage{
kFalling(h0, hf, t, L, As = NA, Ap = NA, Ds = NA, Dp = NA)
}
\arguments{
\item{h0}{= head difference at beginning of test}

\item{hf}{= head difference at end of test (h0 > hf)}

\item{t}{= time of flow}

\item{L}{= length of soil sample}

\item{As}{= cross-sectional area of the soil sample}

\item{Ap}{= cross-sectional area of the standpipe}

\item{Ds}{= diameter of the soil sample}

\item{Dp}{= diameter of the standpipe}
}
\value{
Hydraulic conductivity from the falling head test
}
\description{
Computes the hydraulic conductivity from the falling head test. 1 January
2016
}
\note{
Either English or metric units can be used, but they must be consistent.
Either the areas (As and Ap) OR the diameters (Ds and Dp) need to be specified.
}
\examples{
kFalling(h0 = 12, hf = 2, L = 10, Ds = 20, Dp = 2, t = 100)
# k in units of cm/s

}
\author{
Kyle Elmy and Jim Kaklamanos
}
