% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LateralEarthPressures.R
\encoding{UTF-8}
\name{Ko}
\alias{Ko}
\title{Coefficient of Lateral Earth Pressure at Rest}
\usage{
Ko(phi, OCR = 1)
}
\arguments{
\item{phi}{= effective friction angle (deg)}

\item{OCR}{= overconsolidation ratio (default = 1)}
}
\value{
Coefficient of lateral earth pressure at rest
}
\description{
This function calculates the coefficient of lateral earth pressure at rest.
1 October 2015
}
\note{
For normally consolidated soil, the Jaky (1944) equation is used;
for overconsolidated soil, the Mayne and Kulhawy (1982) equation is used
}
\author{
Kyle Elmy and Jim Kaklamanos
}
