% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LateralEarthPressures.R
\encoding{UTF-8}
\name{Kp}
\alias{Kp}
\title{Rankine Passive Earth Pressures (Lateral Earth Pressure)}
\usage{
Kp(phi, beta = 0)
}
\arguments{
\item{phi}{= effective friction angle (deg)}

\item{beta}{= angle of backfill (deg); default = 0}
}
\value{
Coefficient of lateral earth pressure (passive)
}
\description{
This function calculates the passive coefficient of lateral earth pressure.
1 October 2015
}
\note{
The Rankine theory requires that beta <= phi
}
\examples{
Kp(phi = 30, beta = 10)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
