% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groundwater.R
\encoding{UTF-8}
\name{kPump}
\alias{kPump}
\title{Hydraulic Conductivity From Pumping Tests (Groundwater)}
\usage{
kPump(Q, H = NA, h1, h2, r1, r2)
}
\arguments{
\item{Q}{= flow rate into well}

\item{H}{= thickness of aquifer}

\item{h1}{= total head in farthest observation well}

\item{h2}{= total head in nearest observation well}

\item{r1}{= radius from pumped well to farthest observation well}

\item{r2}{= radius from pumped well to nearest observation well}
}
\value{
k = hydraulic conductivity of aquifer
}
\description{
Computes the hydraulic conductivity from pumping tests. 1 January 2016
}
\note{
Datum for h0 and hf is the bottom of the aquifer.
For unconfined aquifers, H > h1 (or specify as NA)
For confined aquifers, H >= h2.
For mixed aquifers (which start as confined prior to pumping and finish as
unconfined after pumping is complete), H < h2.
}
\examples{
kPump(Q = 14.5, H = 10, h1 = 20, h2 = 16, r1 = 16, r2 = 8)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
