% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groundwater.R
\encoding{UTF-8}
\name{kx}
\alias{kx}
\title{Equivalent Horizontal Hydraulic Conductivity (Groundwater)}
\usage{
kx(thk, k)
}
\arguments{
\item{thk}{= vector of layer thicknesses}

\item{k}{= vector of layer hydraulic conductivities}
}
\value{
Equivalent horizontal hydraulic conductivity (kx) for layered soil
    deposits
}
\description{
Computes the equivalent horizontal hydraulic conductivity. 1 January
2016
}
\author{
Kyle Elmy and Jim Kaklamanos
}
