% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groundwater.R
\encoding{UTF-8}
\name{kz}
\alias{kz}
\title{Equivalent Vertical Hydraulic Conductivity (Groundwater)}
\usage{
kz(thk, k)
}
\arguments{
\item{thk}{= vector of layer thicknesses}

\item{k}{= vector of layer hydraulic conductivities}
}
\value{
Equivalent vertical hydraulic conductivity (kz) for layered soil
    deposits
}
\description{
Computes the equivalent vertical hydraulic conductivity. 1 January
2016
}
