% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plasticity.R
\encoding{UTF-8}
\name{LI}
\alias{LI}
\title{Liquidity Index (Plasticity)}
\usage{
LI(w, LL, PL)
}
\arguments{
\item{w}{= in-situ moisture content (percent)}

\item{LL}{= Liquid Limit (percent)}

\item{PL}{= Plastic Limit (percent)}
}
\value{
A two-element list composed of: LI = Liquidity index (percent) and
   descr = qualitative description of soil in its natural moisture state
}
\description{
Calculates liquidity index from liquid limit, plastic limit, and in-situ
moisture content. 1 December 2015
}
\examples{

LI(w = 55, PL = 20, LL = 50)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
