% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plasticity.R
\encoding{UTF-8}
\name{LL}
\alias{LL}
\title{Liquid Limit (Plasticity)}
\usage{
LL(N, w, draw)
}
\arguments{
\item{N}{= vector of number of blows obtained from the liquid limit test}

\item{w}{= vector of water contents (PERCENT, not decimal)}

\item{draw}{= logical variable: TRUE if plot of flow curve is desired, FALSE
to suppress the creation of the plot}
}
\value{
Liquid limit (percent) and plot of soil's flow curve (optional)
}
\description{
Calculate and plot a soil's liquid limit (LL) using the flow curve. 1
December 2015
}
\examples{
LL(N = c(72, 37, 14), w = c(7, 15, 21), draw = TRUE)
# liquid limit analysis
}
\author{
Kyle Elmy and Jim Kaklamanos
}
