% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBackground.R
\encoding{UTF-8}
\name{logAxis.plot}
\alias{logAxis.plot}
\title{logAxis Plot Background Function}
\usage{
logAxis.plot(data, type, gridlines)
}
\arguments{
\item{data}{= vector of data values for plot;}

\item{type}{= "x" for horizontal axis; "y" for vertical axis}

\item{gridlines}{= vertical TRUE for vertical gridlines across plot;
FALSE otherwise (default = FALSE)}
}
\description{
logAxis.plot plotting function (one axis at a time). 1 December 2015
}
\note{
This file contains functions that are used to create the logarithmic
and semilogarithmic plots that are used by other functions in the package.
}
\author{
Kyle Elmy and Jim Kaklamanos
}
