% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBackground.R
\encoding{UTF-8}
\name{logAxis}
\alias{logAxis}
\title{logAxis Background Wrapper Function for Performing Calculations}
\usage{
logAxis(x = NA, y = NA, gridX = FALSE, gridY = FALSE)
}
\arguments{
\item{x}{= data for x-axis}

\item{y}{= data for y-axis}

\item{gridX}{= TRUE if gridlines in the horizontal directions are to be drawn;
FALSE otherwise}

\item{gridY}{= TRUE if gridlines in the vertical directions are to be drawn;
FALSE otherwise}
}
\description{
Background function for performing calculations. 1 December 2015
}
\author{
Kyle Elmy and Jim Kaklamanos
}
