% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MohrCircle.R
\encoding{UTF-8}
\name{MohrCircle.plot}
\alias{MohrCircle.plot}
\title{Mohr Circle (Plot)}
\usage{
MohrCircle.plot(
  sigmaX = NA,
  sigmaZ = NA,
  tauXZ = NA,
  sigma1 = NA,
  sigma3 = NA,
  metric = TRUE
)
}
\arguments{
\item{sigmaX}{= normal stress acting in the horizontal direction}

\item{sigmaZ}{= normal stress acting in the vertical direction}

\item{tauXZ}{= shear stress acting on the same plane as sigmaX}

\item{sigma1}{= major principal stress}

\item{sigma3}{= minor principal stress}

\item{metric}{= logical variable: TRUE for metric units (kPa), and FALSE for
English units}
}
\value{
Plot of Mohr Circle
}
\description{
This function plots the Mohr circle. 1 October 2015
}
\note{
One of the following two sets of data must be entered:
1.  sigmaX, sigmaZ, tauXZ
2.  sigma1, sigma3
}
\examples{
MohrCircle.plot(sigmaX = 80, sigmaZ = 120, tauXZ = 20, metric = FALSE)
}
\author{
Kyle Elmy and Jim Kaklamanos
}
