% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsurfaceExploration.R
\encoding{UTF-8}
\name{N160}
\alias{N160}
\title{SPT Correction N160 for Site Exploration}
\usage{
N160(N60, sigma, metric)
}
\arguments{
\item{N60}{= SPT N-value corrected for field procedures}

\item{sigma}{= effective vertical stress at the depth of interest (psf or kPa)}

\item{metric}{= logical variable: TRUE (for metric units) or FALSE (for
English units)}
}
\value{
SPT blow count (N-value) corrected for field procedures and
    overburden pressure
}
\description{
This function provides the SPT correction: N1,60. 1 January 2016
}
\examples{

# In English units
N160(N60 = 8, sigma = 1500, metric = FALSE)

# In SI units
N160(N60 = 8, sigma = 90, metric = TRUE) 
 
 
 
}
\author{
Kyle Elmy and Jim Kaklamanos
}
