% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsurfaceExploration.R
\encoding{UTF-8}
\name{N60}
\alias{N60}
\title{SPT Correction N60 for Site Exploration}
\usage{
N60(N, Lr, Db, SS = TRUE, E = 0.6, metric)
}
\arguments{
\item{N}{= raw SPT N-value}

\item{Lr}{= rod length (ft or m)}

\item{Db}{= borehole diameter (in or mm)}

\item{SS}{= logical variable: TRUE for standard sampler [default]; FALSE
for sampler without liner}

\item{E}{= hammer efficiency as a decimal (default: 0.60)}

\item{metric}{= logical variable: TRUE (for metric units) or FALSE (for
English units)}
}
\value{
SPT blow count (N-value) corrected for field procedures
}
\description{
This function provides the SPT correction: N60. 1 January 2016
}
\examples{
N60(N = 11, Lr = 25, Db = 4, E = 0.50, SS = TRUE, metric = FALSE)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
