% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearingCapacity.R
\encoding{UTF-8}
\name{Nq}
\alias{Nq}
\title{Bearing Capacity Factors (Nq)}
\usage{
Nq(phi, case = "general", method = "Terzaghi")
}
\arguments{
\item{phi}{= friction angle (degrees)}

\item{case}{= "general" or "local" to indicate general or local shear
failure ("general" is default)}

\item{method}{= "Terzaghi" or "Vesic" ("Terzaghi" is default)}
}
\value{
Bearing capacity factor (Nq) using either the Terzaghi
   or Vesic methods
}
\description{
This function calculates the bearing capacity factor Nq. 1 October 2015
}
\note{
For local shear, the friction angle is reduced to a value equal to atan(2/3
* tan(phi)).
Terzaghi's Ngamma is approximated by Coduto (2001).
}
\author{
Kyle Elmy and Jim Kaklamanos
}
