% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexParameters.R
\encoding{UTF-8}
\name{phase.params}
\alias{phase.params}
\title{Weight-Volume Parameters (Phase Diagrams and Index Parameters)}
\usage{
phase.params(Ws, Ww, Vs, Vw, Va)
}
\arguments{
\item{Ws}{= weight of solids}

\item{Ww}{= weight of water}

\item{Vs}{= volume of solids}

\item{Vw}{= volume of water}

\item{Va}{= volume of air}
}
\value{
The output is a ten-element list containing: w = water content (as
   decimal), S = degree of saturation (as decimal), e = void ratio (as
   decimal), n = porosity (as decimal), Gs = specific gravity, gammaT = total
   unit weight, gammaD = dry unit weight, gammaS = unit weight of solids,
   gammaW = unit weight of water, and gammaB = buoyant unit weight
}
\description{
This function calculates a comprehensive list of index parameters from
provided weights and volumes of a soil sample. 1 December 2015
}
\examples{
phase.params(Ws = 75.8, Ww = 15.6, Vs = 0.45, Vw = 0.25, Va = 0.1)
}
\author{
Kyle Elmy and Jim Kaklamanos
}
