% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexParameters.R
\encoding{UTF-8}
\name{phase.plot}
\alias{phase.plot}
\title{Phase Diagrams Plot (Phase Diagrams and Index Parameters)}
\usage{
phase.plot(Ws, Ww, Vs, Vw, Va, W.unit, V.unit, mass = FALSE)
}
\arguments{
\item{Ws}{= weight of solids}

\item{Ww}{= weight of water}

\item{Vs}{= volume of solids}

\item{Vw}{= volume of water}

\item{Va}{= volume of air}

\item{W.unit}{= measurement unit of weights}

\item{V.unit}{= measurement unit of volumes}

\item{mass}{= logical variable: TRUE for masses or FALSE for weights (default)}
}
\value{
Plot of phase diagram
}
\description{
This function plots phase diagrams from weights (or masses) and volumes of a
soil sample. 1 December 2015
}
\note{
If any parameters are zero, please enter "0"; do not leave them blank.
Plot is currently not to scale; may be edited in the future to allow this
functionality.
}
\examples{
phase.plot(Ws = 75.8, Ww = 15.6, Vs = 0.45, Vw = 0.25, Va = 0.1, W.unit =
"lb", V.unit = "ft^3", mass = FALSE)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
