% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plasticity.R
\encoding{UTF-8}
\name{PI}
\alias{PI}
\title{Plasticity Index}
\usage{
PI(LL, PL)
}
\arguments{
\item{LL}{= Liquid Limit (percent)}

\item{PL}{= Plastic Limit (percent)}
}
\value{
A two-element list composed of: PI = Plasticity index (percent) and
   descr = qualitative description of soil based on Sowers (1979)
}
\description{
Calculates plasticity index from liquid limit and plastic limit. 1 December
2015
}
\examples{
PI(LL=80, PL=30)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
