% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plasticity.R
\encoding{UTF-8}
\name{plasticity.plot}
\alias{plasticity.plot}
\title{Casagrande's Plasticity Chart}
\usage{
plasticity.plot(LL, PL = NA, PI = NA)
}
\arguments{
\item{LL}{= Liquid Limit (percent)}

\item{PL}{= Plastic Limit (percent)}

\item{PI}{= Plasticity index (percent)}
}
\value{
Plot of a soil's plasticity parameters on Casagrande's plasticity
   chart
}
\description{
This function plots the Casagrande's plasticity chart. 1 December 2015
}
\note{
Either PL or PI must be specified
}
\examples{
plasticity.plot(LL = 40, PL = 20)
}
\author{
Kyle Elmy and Jim Kaklamanos
}
