% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexParameters.R
\encoding{UTF-8}
\name{relDensity}
\alias{relDensity}
\title{Relative Density (Phase Diagrams and Index Parameters)}
\usage{
relDensity(e, emax, emin)
}
\arguments{
\item{e}{= void ratio}

\item{emax}{= maximum void ratio}

\item{emin}{= minimum void ratio}
}
\value{
Relative density as a decimal
}
\description{
Calculates the relative density. 1 December 2015
}
\examples{
relDensity(e = 0.3, emax = 0.92, emin = 0.35)
}
\author{
Kyle Elmy and Jim Kaklamanos
}
