% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MohrCircle.R
\encoding{UTF-8}
\name{sigma13}
\alias{sigma13}
\title{Principal Stresses (Mohr Circle Analyses)}
\usage{
sigma13(sigmaX, sigmaZ, tauXZ)
}
\arguments{
\item{sigmaX}{= normal stress acting in the horizontal direction}

\item{sigmaZ}{= normal stress acting in the vertical direction}

\item{tauXZ}{= shear stress acting on the same plane as sigmaX}
}
\value{
A four-element list containing: sigma1 = magnitude of major
   principal stress, sigma3 = magnitude of minor principal stress, theta1 =
   direction of major principal stress (deg), and theta3 = direction of
   minor principal stress (deg)
}
\description{
This function calculates the principal stresses. 1 October 2015
}
\examples{
sigma13(sigmaX = 80, sigmaZ = 120, tauXZ = 20)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
