% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MohrCircle.R
\encoding{UTF-8}
\name{sigmaTrans}
\alias{sigmaTrans}
\title{Stress Transformation (Mohr Circle Analyses)}
\usage{
sigmaTrans(
  theta,
  sigmaX = NA,
  sigmaZ = NA,
  tauXZ = NA,
  sigma1 = NA,
  sigma3 = NA
)
}
\arguments{
\item{theta}{= angle of inclination (degrees)}

\item{sigmaX}{= normal stress acting in the horizontal direction}

\item{sigmaZ}{= normal stress acting in the vertical direction}

\item{tauXZ}{= shear stress acting on the same plane as sigmaX}

\item{sigma1}{= major principal stress}

\item{sigma3}{= minor principal stress}
}
\value{
A two-element list containing: sigma = normal stress on an inclined
   plane and tau = shear stress on an inclined plane
}
\description{
This function computes the stress transformation. 1 October 2015
}
\note{
In addition to theta, One of the following two sets of data must be entered:
1.  sigmaX, sigmaZ, tauXZ
2.  sigma1, sigma3
If theta is entered in conjunction with sigmaX, sigmaZ, and tauXZ, it is
interpreted as the angle of inclination above the horizontal.  If theta is
entered in conjunction with the principal stresses, then it is interpreted as
the angle of inclination above the major principal plane.
}
\examples{
sigmaTrans(sigmaX = 80, sigmaZ = 120, tauXZ = 20, theta = 78)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
