% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LateralEarthPressures.R
\encoding{UTF-8}
\name{sigmaX.plot}
\alias{sigmaX.plot}
\title{Horizontal Stress Plot (Lateral Earth Pressure)}
\usage{
sigmaX.plot(depth, sigmaX.eff, sigmaX.total = NA, u = NA, metric)
}
\arguments{
\item{depth}{= vector of depths (ft or m)}

\item{sigmaX.eff}{= vector of effective stresses (psf or kPa)}

\item{sigmaX.total}{= vector of total stresses (psf or kPa)}

\item{u}{= vector of pore water pressures (psf or kPa)}

\item{metric}{= logical variable: TRUE (for metric units) or FALSE (for
English units)}
}
\value{
Plot of horizontal stress profile versus depth
}
\description{
This function draws the horizontal stress plot. 1 October 2015
}
\note{
If sigmaX.total and u are left blank, the plot is only constructed for
effective stress
Once constructed, additional profiles may be added to this plot (for
example, for induced stress or maximum past pressure)
}
\examples{
# Site with constant unit weight = 100 pcf, GWT at 10 ft depth
temp <- sigmaX.profile(gamma = rep(100, 3), depth = c(10, 20, 30), K =
c(0.35, 0.30, 0.28), zw = 10, metric = FALSE)                    
depth <- temp$depth
sigmaTotal <- temp$sigmaX.total
u <- temp$u
sigmaEff <- temp$sigmaX.eff
sigmaX.plot(depth = depth, sigmaX.eff = sigmaEff, metric = FALSE,
sigmaX.total = sigmaTotal, u = u)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
