% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LateralEarthPressures.R
\encoding{UTF-8}
\name{sigmaX.profile}
\alias{sigmaX.profile}
\title{Horizontal Stress Profile (Lateral Earth Pressure)}
\usage{
sigmaX.profile(
  gamma,
  thk = NA,
  depth = NA,
  K,
  zw,
  zout = NA,
  gammaW = NA,
  metric
)
}
\arguments{
\item{gamma}{= vector of unit weights (pcf or kN/m^3)}

\item{thk}{= vector of layer thicknesses (ft or m)}

\item{depth}{= vector of layer bottom depths (ft or m)}

\item{K}{= vector of layers' lateral earth pressure coefficients}

\item{zw}{= depth of groundwater table (ft or m)}

\item{zout}{= desired depths of output (ft or m): defaults to critical
locations in the profile (the top and bottom of the profile, layer
interfaces, and the groundwater table) [recommended]}

\item{gammaW}{= unit weight of water (default = 62.4 pcf for English units;
9.81 kN/m^3 for metric units)}

\item{metric}{= logical variable: TRUE (for metric units) or FALSE (for
English units)}
}
\value{
A four-element list containing four vectors: depth = depth,
   sigmaZ.eff = effective vertical stress, sigmaZ.total = total vertical
   stress, and u = pore water pressure
}
\description{
This function calculates the horizontal stress profile. 1 October 2015
}
\examples{
sigmaX.profile(gamma = c(108, 116), depth = c(15, 40), K = c(0.34, 0.32),
zw = 15, metric = FALSE)


}
\author{
Kyle Elmy and Jim Kaklamanos
}
