% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LateralEarthPressures.R
\encoding{UTF-8}
\name{sigmaX}
\alias{sigmaX}
\title{Horizontal Stress at a Point (Lateral Earth Pressure)}
\usage{
sigmaX(
  gamma,
  thk = NA,
  depth = NA,
  zw,
  zout,
  K,
  gammaW = NA,
  metric,
  upper = TRUE
)
}
\arguments{
\item{gamma}{= vector of unit weights (pcf or kN/m^3)}

\item{thk}{= vector of layer thicknesses (ft or m)}

\item{depth}{= vector of layer bottom depths (ft or m)}

\item{zw}{= depth of groundwater table (ft or m)}

\item{zout}{= desired depth of output (ft or m): a single value}

\item{K}{= vector of lateral earth pressure coefficients}

\item{gammaW}{= unit weight of water (default = 62.4 pcf for English units;
9.81 kN/m^3 for metric units)}

\item{metric}{= logical variable: TRUE (for metric units) or FALSE (for
English units)}

\item{upper}{= logical variable to specify whether the upper (TRUE) or lower
(FALSE) lateral earth pressure coefficient should be used, for the special
case that zout corresponds to a layer interface}
}
\value{
A three-element list containing: sigmaX.eff = effective horizontal
   stress, sigmaX.total = total horizontal stress, and u = pore water pressure
}
\description{
This function calculates the horizontal stress at a point. 1 October 2015
}
\examples{
sigmaX(gamma = c(108, 116), depth = c(15, 40), zout = 18, K = c(0.34, 0.32),
zw = 15, metric = FALSE, upper = TRUE)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
