% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stress.R
\encoding{UTF-8}
\name{sigmaZ.plot}
\alias{sigmaZ.plot}
\title{Vertical Stress Plot}
\usage{
sigmaZ.plot(depth, sigmaZ.eff, sigmaZ.total = NA, u = NA, metric)
}
\arguments{
\item{depth}{= vector of depths (ft or m)}

\item{sigmaZ.eff}{= vector of effective stresses (psf or kPa)}

\item{sigmaZ.total}{= vector of total stresses (psf or kPa)}

\item{u}{= vector of pore water pressures (psf or kPa)}

\item{metric}{= logical variable: TRUE (for metric units) or FALSE (for
English units)}
}
\value{
Plot of vertical stress versus depth
}
\description{
This function draws the vertical stress plot. 1 October 2015
}
\note{
If sigmaZ.total and u are left blank, the plot is only constructed for
effective stress
Once constructed, additional profiles may be added to this plot
(for example, for induced stress or maximum past pressure)
}
\examples{
# Site with constant unit weight = 100 pcf, GWT at 10 ft depth
temp <- sigmaZ.profile(gamma = rep(100, 3), depth = c(10, 20, 30), zw = 10,
metric = FALSE)
depth <- temp$depth
sigmaTotal <- temp$sigmaZ.total
u <- temp$u
sigmaEff <- temp$sigmaZ.eff
sigmaZ.plot(depth = depth, sigmaZ.eff = sigmaEff,
metric = FALSE, sigmaZ.total = sigmaTotal, u = u)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
