% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stress.R
\encoding{UTF-8}
\name{sigmaZ.profile}
\alias{sigmaZ.profile}
\title{Vertical Stress Profile}
\usage{
sigmaZ.profile(gamma, thk = NA, depth = NA, zw, zout = NA, gammaW = NA, metric)
}
\arguments{
\item{gamma}{= vector of unit weights (pcf or kN/m^3)}

\item{thk}{= vector of layer thicknesses (ft or m)}

\item{depth}{= vector of layer bottom depths (ft or m)}

\item{zw}{= depth of groundwater table (ft or m)}

\item{zout}{= desired depths of output (ft or m): defaults to critical
locations in the profile (the top and bottom of the profile, layer
interfaces, and the groundwater table)  [recommended]}

\item{gammaW}{= unit weight of water (default = 62.4 pcf for English units;
9.81 kN/m^3 for metric units)}

\item{metric}{= logical variable: TRUE (for metric units) or FALSE (for
English units)}
}
\value{
A four-element list containing four vectors: depth = depth,
   sigmaZ.eff = effective vertical stress, sigmaZ.total = total vertical
   stress, and u = pore water pressure
}
\description{
This function computes the vertical stress profile. 1 October 2015
}
\examples{
sigmaZ.profile(gamma = c(108, 116), depth = c(15, 40), zw = 15, metric =
FALSE)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
