% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stress.R
\encoding{UTF-8}
\name{sigmaZ}
\alias{sigmaZ}
\title{Vertical Stress at a Point}
\usage{
sigmaZ(gamma, thk = NA, depth = NA, zw, zout, gammaW = NA, metric)
}
\arguments{
\item{gamma}{= vector of unit weights (pcf or kN/m^3)}

\item{thk}{= vector of layer thicknesses (ft or m)}

\item{depth}{= vector of layer bottom depths (ft or m)}

\item{zw}{= depth of groundwater table (ft or m)}

\item{zout}{= desired depth of output (ft or m): a single value}

\item{gammaW}{= unit weight of water (default = 62.4 pcf for English units;
9.81 kN/m^3 for metric units)}

\item{metric}{= logical variable: TRUE (for metric units) or FALSE (for
English units)}
}
\value{
A three-element list containing: sigmaZ.eff = effective vertical
    stress, sigmaZ.total = total vertical stress, and u = pore water pressure
}
\description{
This function computes the vertical stress at a point. 1 October 2015
}
\note{
Either layer thicknesses or depths to layer bottoms must be specified.
}
\examples{
sigmaZ(gamma = c(108, 116), depth = c(15, 40), zout = 18, zw = 15, metric =
FALSE)
}
\author{
Kyle Elmy and Jim Kaklamanos
}
