% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrainSize.R
\encoding{UTF-8}
\name{size.from.sieve}
\alias{size.from.sieve}
\title{Grain Sizes Of Sieves}
\usage{
size.from.sieve(sieve, metric)
}
\arguments{
\item{sieve}{= vector of sieve numbers according to ASTM D422}

\item{metric}{= logical variable: TRUE for metric units (mm), FALSE for
English units (in)}
}
\value{
Vector of grain sizes (in or mm)
}
\description{
Calculate a set of grain sizes corresponding to a set of sieves. 1 December
2015
}
\note{
For sieves larger than the no. 4 sieve, the user should specify the sieve
size in inches (e.g., 3/8, 3/4, 1, 1.5, 2, 3, etc.)
Either sieve numbers OR grain sizes must be provided.
}
\author{
Kyle Elmy and Jim Kaklamanos
}
