% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsurfaceExploration.R
\encoding{UTF-8}
\name{soil.profile}
\alias{soil.profile}
\title{Soil Profile Plot for Site Exploration}
\usage{
soil.profile(
  thk = NA,
  depth = NA,
  zw,
  type = NA,
  gamma = NA,
  phi = NA,
  C = NA,
  title = "Soil Profile",
  metric
)
}
\arguments{
\item{thk}{= vector of layer thicknesses (ft or m)}

\item{depth}{= vector of layer bottom depths (ft or m)}

\item{zw}{= depth of groundwater table (ft or m)}

\item{type}{= vector of soil types (character strings)}

\item{gamma}{= vector of unit weights (pcf or kN/m^3)}

\item{phi}{= vector of soil friction angles (deg)}

\item{C}{= vector of soil cohesion (psf or kPa)}

\item{title}{= desired title of plot (default: "Soil Profile")}

\item{metric}{= logical variable: TRUE (for metric units) or FALSE (for
English units)}
}
\value{
A plot of a soil profile
}
\description{
Plots the soil profile for use in site exploration. 1 January 2016
}
\note{
Either layer thicknesses or depth to layer bottoms must be specified.
The only necessary variables are thk (or depth), zw, and metric.  All other
variables are optional.
}
\examples{
soil.profile(depth = c(20, 40, 52, 60), zw = 20,
             type = c("Dry Sand", "Saturated Sand", "Soft Clay",
               "Dense Gravel"),
             gamma = c(110, 115, 120, 150), phi = c(30, 30, NA, 38),
             C = c(NA, NA, 300, NA), metric = FALSE)


}
\author{
Kyle Elmy and Jim Kaklamanos
}
