% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MohrCircle.R
\encoding{UTF-8}
\name{tauMax}
\alias{tauMax}
\title{Maximum In-Plane Shear Stress (Mohr Circle Analyses)}
\usage{
tauMax(sigmaX, sigmaZ, tauXZ)
}
\arguments{
\item{sigmaX}{= normal stress acting in the horizontal direction}

\item{sigmaZ}{= normal stress acting in the vertical direction}

\item{tauXZ}{= shear stress acting on the same plane as sigmaX}
}
\value{
A two-element list containing: tauMax = maximum in-plane shear stress
   and theta = angle of maximum in-plane shear stress
}
\description{
This function calculates the maximum in-plane shear stress. 1 October 2015
}
\examples{
tauMax(sigmaX = 80, sigmaZ = 120, tauXZ = 20)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
