% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilClassification.R
\encoding{UTF-8}
\name{USCS.coarse.symbol}
\alias{USCS.coarse.symbol}
\title{USCS Soil Classification (USCS Coarse-Grained Symbol)}
\usage{
USCS.coarse.symbol(pg, ps, pf, Cc, Cu, PI = NA, LL, PL)
}
\arguments{
\item{pg}{= percent gravel (\%)}

\item{ps}{= percent sand (\%)}

\item{pf}{= percent fines (\%)}

\item{Cc}{= coefficient of curvature}

\item{Cu}{= coefficient of uniformity}

\item{PI}{= plasticity index}

\item{LL}{= liquid limit}

\item{PL}{= plastic limit}
}
\value{
A soil's USCS coarse-grained symbol
}
\description{
This function classifies soil using the Unified Soil Classification
System (USCS) using the coarse-grained symbol. 1 December 2015
}
\note{
Data on the soil's gradation (Cu and Cc) are required if the percent fines
is less than or equal to 12%. Data on the soil's fines [either (a) LL and PL
or (b) LL and PI] are required if the percent fines is greater than or equal
to 5%.
}
\author{
Kyle Elmy and Jim Kaklamanos
}
