% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilClassification.R
\encoding{UTF-8}
\name{USCS.fine.symbol}
\alias{USCS.fine.symbol}
\title{USCS Soil Classification (USCS Fine-Grained Symbol)}
\usage{
USCS.fine.symbol(LL, PL = NA, PI = NA)
}
\arguments{
\item{LL}{= Liquid Limit}

\item{PL}{= Plastic Limit}

\item{PI}{= Plasticity Index}
}
\value{
A soil's fine-grained USCS symbol
}
\description{
This function classifies soil using the Unified Soil Classification
System (USCS) using the fine-grained symbol. 1 December 2015
}
\note{
Either PL or PI must be provided.
}
\author{
Kyle Elmy and Jim Kaklamanos
}
