% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilClassification.R
\encoding{UTF-8}
\name{USCS}
\alias{USCS}
\title{USCS Soil Classification}
\usage{
USCS(
  pg = NA,
  ps = NA,
  pf = NA,
  Cc = NA,
  Cu = NA,
  LL = NA,
  PL = NA,
  PI = NA,
  sieve = NA,
  size = NA,
  percent = NA,
  metric = NA
)
}
\arguments{
\item{pg}{= percent gravel (\%)}

\item{ps}{= percent sand (\%)}

\item{pf}{= percent fines (\%)}

\item{Cc}{= coefficient of curvature}

\item{Cu}{= coefficient of uniformity}

\item{LL}{= liquid limit}

\item{PL}{= plastic limit}

\item{PI}{= plasticity index}

\item{sieve}{= vector of sieve numbers according to ASTM D422 for grain-size
distribution}

\item{size}{= vector of grain sizes (in or mm) in distribution}

\item{percent}{= vector of percent passing (grain-size distribution)}

\item{metric}{= logical variable: TRUE for metric units (mm), FALSE for
English units (in) [only needed if grain sizes are provided]}
}
\value{
A two-element list providing a soil's USCS group symbol and name
   symbol = USCS group symbol and name = USCS group name
}
\description{
This is a master function that calculates a soil's classification according
to the Unified Soil Classification System (USCS). The functions A.line,
USCS.coarse.symbol, and USCS.fine.symbol are called by this function. 1
December 2015
}
\note{
Data on the soil's grain-size distribution are required if the percent fines
is less than or equal to 12%. The user has three options for input to this
function:          
1.  Sieve numbers (sieve); and percent passing
2.  Grain sizes (size); and percent passing
3.  Percent components (pg, ps, pf) and coefficients of uniformity and
curvature (Cc and Cu)           
Data on the soil's fines [either (a) LL and PL or (b) LL and PI] are required
if the percent fines is greater than or equal to 5%.
If percent components are specified, then Cu and Cc are needed if pf < 85%.
}
\examples{
USCS(pg = 15, ps = 34, pf = 51, Cc = 1, Cu = 4, LL = 40, PL = 10)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
