% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexParameters.R
\encoding{UTF-8}
\name{waterContent}
\alias{waterContent}
\title{Water Content (Phase Diagrams and Index Parameters)}
\usage{
waterContent(M1, M2, Mc)
}
\arguments{
\item{M1}{= Mass (or weight) of can and wet soil, before oven dry}

\item{M2}{= Mass (or weight) of can and dry soil, after drying in oven}

\item{Mc}{= Mass (or weight) of can}
}
\value{
Water content as a decimal
}
\description{
Calculates water content from lab results (i.e., measured weights or masses)
. 1 December 2015
}
\note{
Either mass or weight can be used, since the units cancel.
}
\examples{
waterContent(M1 = 20.68, M2 = 18.14, Mc = 8.20)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
