% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groundwater.R
\encoding{UTF-8}
\name{wellDrawdown}
\alias{wellDrawdown}
\title{Well Drawdown (Groundwater)}
\usage{
wellDrawdown(Q, k, H = NA, h0, r0, rw, r)
}
\arguments{
\item{Q}{= flow rate into well}

\item{k}{= hydraulic conductivity of aquifer}

\item{H}{= saturated thickness of aquifer}

\item{h0}{= initial total head in aquifer (before pumping)}

\item{r0}{= radius of influence}

\item{rw}{= radius of well}

\item{r}{= radius of interest}
}
\value{
A two-element list containing: h = height of groundwater surface a
   distance r from the well and dd = h0 - h = drawdown of groundwater
   surface a distance r from the well
}
\description{
Computes the well drawdown. 1 January 2016
}
\note{
Datum for total heads are the bottom of the aquifer.
For unconfined aquifers, H = NA.
For confined aquifers, H >= hf.
For mixed aquifers (which start as confined prior to pumping and finish as
unconfined after pumping is complete), H < hf.
}
\examples{
wellDrawdown(Q = 14.5, k = 0.065, H = 10, h0 = 21, r0 = 20, r = 2, rw = 2)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
