% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groundwater.R
\encoding{UTF-8}
\name{wellFlow}
\alias{wellFlow}
\title{Flow Rate to Wells (Groundwater)}
\usage{
wellFlow(k, H = NA, h0, hf, r0, rw)
}
\arguments{
\item{k}{= hydraulic conductivity of aquifer}

\item{H}{= thickness of aquifer}

\item{h0}{= initial total head in aquifer (before pumping)}

\item{hf}{= final total head in well casing (after pumping)}

\item{r0}{= radius of influence}

\item{rw}{= radius of well}
}
\value{
Q = flow rate to well
}
\description{
Computes the flow rate to wells. 1 January 2016
}
\note{
Datum for h0 and hf is the bottom of the aquifer.
For unconfined aquifers, H > h0 (or specify as NA)
For confined aquifers, H >= hf.
For mixed aquifers (which start as confined prior to pumping and finish as
unconfined after pumping is complete), H < hf.
}
\examples{
wellFlow(k = 0.065, H = 10, h0 = 21, hf = 15, r0 = 20, rw = 2)

}
\author{
Kyle Elmy and Jim Kaklamanos
}
